refactor(domain): update all usecases for sync support
