+++ app/src/main/java/com/example/shmr_finance_app_android/domain/usecases/CreateTransactionUseCase.kt
@@ -1,8 +1,7 @@
 package com.example.shmr_finance_app_android.domain.usecases
 
+import com.example.shmr_finance_app_android.core.network.AppError
 import com.example.shmr_finance_app_android.core.utils.combineDateTimeToIso
-import com.example.shmr_finance_app_android.data.remote.api.AppError
-import com.example.shmr_finance_app_android.data.remote.api.NetworkChecker
 import com.example.shmr_finance_app_android.domain.model.TransactionDomain
 import com.example.shmr_finance_app_android.domain.model.TransactionResponseDomain
 import com.example.shmr_finance_app_android.domain.repository.TransactionsRepository
@@ -16,8 +15,7 @@ import javax.inject.Inject
  */
 @Reusable
 class CreateTransactionUseCase @Inject constructor(
-    private val repository: TransactionsRepository,
-    private val networkChecker: NetworkChecker
+    private val repository: TransactionsRepository
 ) {
     /**
      * Запускает UseCase через operator invoke().
@@ -40,10 +38,6 @@ class CreateTransactionUseCase @Inject constructor(
         transactionTime: String,
         comment: String
     ): Result<TransactionDomain> {
-        if (!networkChecker.isNetworkAvailable()) {
-            return Result.failure(AppError.Network)
-        }
-
         return repository.createTransaction(
             accountId,
             categoryId,
