+++ app/src/main/java/com/example/shmr_finance_app_android/domain/usecases/DeleteTransactionUseCase.kt
@@ -1,7 +1,6 @@
 package com.example.shmr_finance_app_android.domain.usecases
 
-import com.example.shmr_finance_app_android.data.remote.api.AppError
-import com.example.shmr_finance_app_android.data.remote.api.NetworkChecker
+import com.example.shmr_finance_app_android.core.network.AppError
 import com.example.shmr_finance_app_android.domain.repository.TransactionsRepository
 import dagger.Reusable
 import javax.inject.Inject
@@ -13,8 +12,7 @@ import javax.inject.Inject
  */
 @Reusable
 class DeleteTransactionUseCase @Inject constructor(
-    private val repository: TransactionsRepository,
-    private val networkChecker: NetworkChecker
+    private val repository: TransactionsRepository
 ) {
     /**
      * Запускает UseCase через operator invoke().
@@ -25,10 +23,6 @@ class DeleteTransactionUseCase @Inject constructor(
      * ошибками из [TransactionsRepository] - при проблемах запроса
      */
     suspend operator fun invoke(transactionId: Int): Result<Unit> {
-        if (!networkChecker.isNetworkAvailable()) {
-            return Result.failure(AppError.Network)
-        }
-
         return repository.deleteTransactionById(transactionId)
     }
 }
