+++ app/src/main/java/com/example/shmr_finance_app_android/domain/usecases/GetAccountUseCase.kt
@@ -1,7 +1,7 @@
 package com.example.shmr_finance_app_android.domain.usecases
 
-import com.example.shmr_finance_app_android.data.remote.api.AppError
-import com.example.shmr_finance_app_android.data.remote.api.NetworkChecker
+import com.example.shmr_finance_app_android.core.network.AppError
+import com.example.shmr_finance_app_android.core.network.NetworkChecker
 import com.example.shmr_finance_app_android.domain.model.AccountResponseDomain
 import com.example.shmr_finance_app_android.domain.repository.AccountRepository
 import dagger.Reusable
@@ -16,8 +16,7 @@ import javax.inject.Inject
  */
 @Reusable
 class GetAccountUseCase @Inject constructor(
-    private val repository: AccountRepository,
-    private val networkChecker: NetworkChecker
+    private val repository: AccountRepository
 ) {
     /**
      * Запускает UseCase через operator invoke().
@@ -28,10 +27,6 @@ class GetAccountUseCase @Inject constructor(
      * ошибками из [AccountRepository] - при проблемах запроса
      */
     suspend operator fun invoke(accountId: Int): Result<AccountResponseDomain> {
-        if (!networkChecker.isNetworkAvailable()) {
-            return Result.failure(AppError.Network)
-        }
-
         return repository.getAccountById(accountId)
     }
 }
