+++ app/src/main/java/com/example/shmr_finance_app_android/domain/usecases/GetTransactionsByPeriodUseCase.kt
@@ -1,7 +1,7 @@
 package com.example.shmr_finance_app_android.domain.usecases
 
-import com.example.shmr_finance_app_android.data.remote.api.AppError
-import com.example.shmr_finance_app_android.data.remote.api.NetworkChecker
+import com.example.shmr_finance_app_android.core.network.AppError
+import com.example.shmr_finance_app_android.core.network.NetworkChecker
 import com.example.shmr_finance_app_android.domain.model.TransactionResponseDomain
 import com.example.shmr_finance_app_android.domain.repository.TransactionsRepository
 import dagger.Reusable
@@ -16,8 +16,7 @@ import javax.inject.Inject
  */
 @Reusable
 class GetTransactionsByPeriodUseCase @Inject constructor(
-    private val repository: TransactionsRepository,
-    private val networkChecker: NetworkChecker
+    private val repository: TransactionsRepository
 ) {
     /**
      * Запускает UseCase через operator invoke().
@@ -36,10 +35,6 @@ class GetTransactionsByPeriodUseCase @Inject constructor(
         startDate: String? = null,
         endDate: String? = null
     ): Result<List<TransactionResponseDomain>> {
-        if (!networkChecker.isNetworkAvailable()) {
-            return Result.failure(AppError.Network)
-        }
-
         return repository.getTransactionsByPeriod(accountId, startDate, endDate)
             .map { list ->
                 list.sortedByDescending { it.transactionDate }
