+++ app/src/main/java/com/example/shmr_finance_app_android/domain/usecases/UpdateTransactionUseCase.kt
@@ -1,8 +1,7 @@
 package com.example.shmr_finance_app_android.domain.usecases
 
+import com.example.shmr_finance_app_android.core.network.AppError
 import com.example.shmr_finance_app_android.core.utils.combineDateTimeToIso
-import com.example.shmr_finance_app_android.data.remote.api.AppError
-import com.example.shmr_finance_app_android.data.remote.api.NetworkChecker
 import com.example.shmr_finance_app_android.domain.model.TransactionResponseDomain
 import com.example.shmr_finance_app_android.domain.repository.TransactionsRepository
 import dagger.Reusable
@@ -15,8 +14,7 @@ import javax.inject.Inject
  */
 @Reusable
 class UpdateTransactionUseCase @Inject constructor(
-    private val repository: TransactionsRepository,
-    private val networkChecker: NetworkChecker
+    private val repository: TransactionsRepository
 ) {
     /**
      * Запускает UseCase через operator invoke().
@@ -41,10 +39,6 @@ class UpdateTransactionUseCase @Inject constructor(
         transactionTime: String,
         comment: String
     ): Result<TransactionResponseDomain> {
-        if (!networkChecker.isNetworkAvailable()) {
-            return Result.failure(AppError.Network)
-        }
-
         return repository.updateTransactionById(
             transactionId = transactionId,
             accountId = accountId,
